requirements.txt
antlr4-python3-runtime==4.7.2
attrs==24.3.0
bidict==0.22.1
filelock==3.16.1
funcy==1.18
lifted-pddl==1.2.3
multipledispatch==1.0.0
numpy==1.24.4
pandas==2.0.3
ply==3.11
psutil==6.1.1
py-aiger==6.2.3
py-aiger-cnf==5.0.8
pyapproxmc==4.1.24
pycryptosat==5.11.21
pypblib==0.0.4
pyrddl==0.1.10
pyrsistent==0.19.3
python-dateutil==2.9.0
python-sat==1.8.dev13
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sortedcontainers==2.4.0
tarski==0.8.2
typing==3.7.4.3